Done date

+++ cpp09/ex00/BitcoinExchange.hpp
@@ -2,13 +2,13 @@
 #include <fstream>
 #include <string>
 #include <map>
-#include "Date.hpp"
+#include "DateConvertor.hpp"
 
 class BitcoinExchange
 {
 	private:
 		const std::string _data_file_path;
-		const std::map<unsigned long, unsigned int> _data_base;
+		const std::map<unsigned long, double> _data_base;
 
 	public:
 		BitcoinExchange(std::string input);
